requirements.txt
fastapi==0.104.1
Jinja2==3.1.2
PyYAML==6.0.1
pandas==2.1.2
uvicorn==0.23.2